+++ src/App.js
@@ -1,0 +1,29 @@
+import React, { Component } from "react";
+import { StyleSheet, View } from "react-native";
+import AppNavigator from "./navigation/AppNavigator";
+
+export default class App extends Component {
+  render() {
+    return (
+      <View style={styles.container}>
+        <AppNavigator />
+      </View>
+    );
+  }
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: "#F5FCFF",
+  },
+  welcome: {
+    fontSize: 20,
+    textAlign: "center",
+  },
+  instructions: {
+    textAlign: "center",
+    color: "#333333",
+    marginBottom: 5,
+  },
+});
